+++ tiled/_tests/adapters/test_sql.py
@@ -6,7 +6,12 @@ import adbc_driver_sqlite
 import pyarrow as pa
 import pytest
 
-from tiled.adapters.sql import SQLAdapter, check_table_name
+from tiled.adapters.sql import (
+    COLUMN_NAME_PATTERN,
+    TABLE_NAME_PATTERN,
+    SQLAdapter,
+    is_safe_identifier,
+)
 from tiled.storage import parse_storage, register_storage
 from tiled.structures.core import StructureFamily
 from tiled.structures.data_source import DataSource, Management
@@ -54,9 +59,7 @@ def data_source_from_init_storage() -> Callable[[str, int], DataSource[TableStru
 
         storage = parse_storage(data_uri)
         register_storage(storage)
-        return SQLAdapter.init_storage(
-            data_source=data_source, storage=storage, path_parts=[]
-        )
+        return SQLAdapter.init_storage(data_source=data_source, storage=storage)
 
     return _data_source_from_init_storage
 
@@ -65,9 +68,9 @@ def data_source_from_init_storage() -> Callable[[str, int], DataSource[TableStru
 def adapter_duckdb_one_partition(
     tmp_path: Path,
     data_source_from_init_storage: Callable[[str, int], DataSource[TableStructure]],
+    duckdb_uri: str,
 ) -> Generator[SQLAdapter, None, None]:
-    data_uri = f"duckdb:///{tmp_path}/test.db"
-    data_source = data_source_from_init_storage(data_uri, 1)
+    data_source = data_source_from_init_storage(duckdb_uri, 1)
     yield SQLAdapter(
         data_source.assets[0].data_uri,
         data_source.structure,
@@ -77,12 +80,12 @@ def adapter_duckdb_one_partition(
 
 
 @pytest.fixture
-def adapter_duckdb_many_partition(
+def adapter_duckdb_many_partitions(
     tmp_path: Path,
     data_source_from_init_storage: Callable[[str, int], DataSource[TableStructure]],
+    duckdb_uri: str,
 ) -> Generator[SQLAdapter, None, None]:
-    data_uri = f"duckdb:///{tmp_path}/test.db"
-    data_source = data_source_from_init_storage(data_uri, 3)
+    data_source = data_source_from_init_storage(duckdb_uri, 3)
     yield SQLAdapter(
         data_source.assets[0].data_uri,
         data_source.structure,
@@ -99,11 +102,13 @@ def test_attributes_duckdb_one_part(adapter_duckdb_one_partition: SQLAdapter) ->
     )
 
 
-def test_attributes_duckdb_many_part(adapter_duckdb_many_partition: SQLAdapter) -> None:
-    assert adapter_duckdb_many_partition.structure().columns == names
-    assert adapter_duckdb_many_partition.structure().npartitions == 3
+def test_attributes_duckdb_many_part(
+    adapter_duckdb_many_partitions: SQLAdapter,
+) -> None:
+    assert adapter_duckdb_many_partitions.structure().columns == names
+    assert adapter_duckdb_many_partitions.structure().npartitions == 3
     assert isinstance(
-        adapter_duckdb_many_partition.conn, adbc_driver_duckdb.dbapi.Connection
+        adapter_duckdb_many_partitions.conn, adbc_driver_duckdb.dbapi.Connection
     )
 
 
@@ -111,9 +116,9 @@ def test_attributes_duckdb_many_part(adapter_duckdb_many_partition: SQLAdapter) 
 def adapter_sql_one_partition(
     tmp_path: Path,
     data_source_from_init_storage: Callable[[str, int], DataSource[TableStructure]],
+    sqlite_uri: str,
 ) -> Generator[SQLAdapter, None, None]:
-    data_uri = f"sqlite:///{tmp_path}/test.db"
-    data_source = data_source_from_init_storage(data_uri, 1)
+    data_source = data_source_from_init_storage(sqlite_uri, 1)
     yield SQLAdapter(
         data_source.assets[0].data_uri,
         data_source.structure,
@@ -123,12 +128,12 @@ def adapter_sql_one_partition(
 
 
 @pytest.fixture
-def adapter_sql_many_partition(
+def adapter_sql_many_partitions(
     tmp_path: Path,
     data_source_from_init_storage: Callable[[str, int], DataSource[TableStructure]],
+    sqlite_uri: str,
 ) -> Generator[SQLAdapter, None, None]:
-    data_uri = f"sqlite:///{tmp_path}/test.db"
-    data_source = data_source_from_init_storage(data_uri, 3)
+    data_source = data_source_from_init_storage(sqlite_uri, 3)
     yield SQLAdapter(
         data_source.assets[0].data_uri,
         data_source.structure,
@@ -145,20 +150,20 @@ def test_attributes_sql_one_part(adapter_sql_one_partition: SQLAdapter) -> None:
     )
 
 
-def test_attributes_sql_many_part(adapter_sql_many_partition: SQLAdapter) -> None:
-    assert adapter_sql_many_partition.structure().columns == names
-    assert adapter_sql_many_partition.structure().npartitions == 3
+def test_attributes_sql_many_part(adapter_sql_many_partitions: SQLAdapter) -> None:
+    assert adapter_sql_many_partitions.structure().columns == names
+    assert adapter_sql_many_partitions.structure().npartitions == 3
     assert isinstance(
-        adapter_sql_many_partition.conn, adbc_driver_sqlite.dbapi.Connection
+        adapter_sql_many_partitions.conn, adbc_driver_sqlite.dbapi.Connection
     )
 
 
 @pytest.fixture
 def adapter_psql_one_partition(
     data_source_from_init_storage: Callable[[str, int], DataSource[TableStructure]],
-    postgresql_database_uri: str,
+    postgres_uri: str,
 ) -> Generator[SQLAdapter, None, None]:
-    data_source = data_source_from_init_storage(postgresql_database_uri, 1)
+    data_source = data_source_from_init_storage(postgres_uri, 1)
     adapter = SQLAdapter(
         data_source.assets[0].data_uri,
         data_source.structure,
@@ -170,11 +175,11 @@ def adapter_psql_one_partition(
 
 
 @pytest.fixture
-def adapter_psql_many_partition(
+def adapter_psql_many_partitions(
     data_source_from_init_storage: Callable[[str, int], DataSource[TableStructure]],
-    postgresql_database_uri: str,
+    postgres_uri: str,
 ) -> SQLAdapter:
-    data_source = data_source_from_init_storage(postgresql_database_uri, 3)
+    data_source = data_source_from_init_storage(postgres_uri, 3)
     return SQLAdapter(
         data_source.assets[0].data_uri,
         data_source.structure,
@@ -283,9 +288,9 @@ def assert_same_rows(table1: pa.Table, table2: pa.Table) -> None:
 @pytest.mark.parametrize(
     "adapter",
     [
-        ("adapter_sql_many_partition"),
-        ("adapter_duckdb_many_partition"),
-        ("adapter_psql_many_partition"),
+        ("adapter_sql_many_partitions"),
+        ("adapter_duckdb_many_partitions"),
+        ("adapter_psql_many_partitions"),
     ],
 )
 def test_append_single_partition(
@@ -311,8 +316,8 @@ def test_append_single_partition(
 @pytest.mark.parametrize(
     "adapter",
     [
-        ("adapter_sql_many_partition"),
-        ("adapter_psql_many_partition"),
+        ("adapter_sql_many_partitions"),
+        ("adapter_psql_many_partitions"),
     ],
 )
 def test_write_read_one_batch_many_part(
@@ -393,112 +398,309 @@ def test_write_read_one_batch_many_part(
         (
             "table_abcdefg12423pnjsbldfhjdfbv_hbdhfljb128w40_ndgjfsdflfnscljm",
             pytest.raises(
-                ValueError, match="Table name is too long, max character number is 63!"
+                ValueError, match=r"Invalid SQL identifier.+max character number is 63"
             ),
         ),
         (
             "create_abcdefg12423pnjsbldfhjdfbv_hbdhfljb128w40_ndgjfsdflfnscljk_sdbf_jhvjkbefl",
             pytest.raises(
-                ValueError, match="Table name is too long, max character number is 63!"
+                ValueError, match=r"Invalid SQL identifier.+max character number is 63"
             ),
         ),
         (
             "hello_abcdefg12423pnjsbldfhjdfbv_hbdhfljb128w40_ndgjfsdflfnscljk_sdbf_jhvjkbefl",
             pytest.raises(
-                ValueError, match="Table name is too long, max character number is 63!"
+                ValueError, match=r"Invalid SQL identifier.+max character number is 63"
             ),
         ),
         ("my_table_here_123_", None),
         ("the_short_table12374620_hello_table23704ynnm", None),
-    ],
-)
-def test_check_table_name_long_name(
-    table_name: str, expected: Union[None, Any]
-) -> None:
-    if isinstance(expected, type(pytest.raises(ValueError))):
-        with expected:
-            check_table_name(table_name)
-    else:
-        assert check_table_name(table_name) is None  # type: ignore[func-returns-value]
-
-
-@pytest.mark.parametrize(
-    "table_name, expected",
-    [
         (
             "_here_is_my_table",
-            pytest.raises(ValueError, match="Illegal table name!"),
+            pytest.raises(ValueError, match=r"Malformed SQL identifier.+"),
+        ),
+        (
+            "here-is-my-table",
+            pytest.raises(ValueError, match=r"Malformed SQL identifier.+"),
+        ),
+        (
+            "-here-is-my-table",
+            pytest.raises(ValueError, match=r"Malformed SQL identifier.+"),
+        ),
+        (
+            "here-is-my-table-",
+            pytest.raises(ValueError, match=r"Malformed SQL identifier.+"),
         ),
         (
             "create_this_table1246*",
-            pytest.raises(ValueError, match="Illegal table name!"),
+            pytest.raises(ValueError, match=r"Malformed SQL identifier.+"),
         ),
         (
             "create this_table1246",
-            pytest.raises(ValueError, match="Illegal table name!"),
+            pytest.raises(ValueError, match=r"Malformed SQL identifier.+"),
         ),
         (
             "drop this_table1246",
-            pytest.raises(ValueError, match="Illegal table name!"),
+            pytest.raises(ValueError, match=r"Malformed SQL identifier.+"),
         ),
         (
             "table_mytable!",
-            pytest.raises(ValueError, match="Illegal table name!"),
+            pytest.raises(ValueError, match=r"Malformed SQL identifier.+"),
         ),
         ("my_table_here_123_", None),
         ("the_short_table12374620_hello_table23704ynnm", None),
-    ],
-)
-def test_check_table_name_illegal_name(
-    table_name: str, expected: Union[None, Any]
-) -> None:
-    if isinstance(expected, type(pytest.raises(ValueError))):
-        with expected:
-            check_table_name(table_name)
-    else:
-        assert check_table_name(table_name) is None  # type: ignore[func-returns-value]
-
-
-@pytest.mark.parametrize(
-    "table_name, expected",
-    [
         (
             "select",
             pytest.raises(
                 ValueError,
-                match="Reserved SQL keywords are not allowed in the table name!",
+                match=r"Reserved SQL keywords are not allowed in identifiers.+",
             ),
         ),
         (
             "create",
             pytest.raises(
                 ValueError,
-                match="Reserved SQL keywords are not allowed in the table name!",
+                match=r"Reserved SQL keywords are not allowed in identifiers.+",
             ),
         ),
         (
             "SELECT",
             pytest.raises(
                 ValueError,
-                match="Reserved SQL keywords are not allowed in the table name!",
+                match=r"Reserved SQL keywords are not allowed in identifiers.+",
             ),
         ),
         (
             "from",
             pytest.raises(
                 ValueError,
-                match="Reserved SQL keywords are not allowed in the table name!",
+                match=r"Reserved SQL keywords are not allowed in identifiers.+",
             ),
         ),
         ("drop_this_table123_", None),
-        ("DROP_thistable123_hwejk", None),
+        (
+            "DROP_thistable123_hwejk",
+            pytest.raises(ValueError, match=r"Malformed SQL identifier.+"),
+        ),
+        (
+            "CAPITALIZED_NAME",
+            pytest.raises(ValueError, match=r"Malformed SQL identifier.+"),
+        ),
     ],
 )
-def test_check_table_name_reserved_keywords(
-    table_name: str, expected: Union[None, Any]
-) -> None:
+def test_check_table_name_is_safe(table_name: str, expected: Union[None, Any]) -> None:
     if isinstance(expected, type(pytest.raises(ValueError))):
         with expected:
-            check_table_name(table_name)
+            is_safe_identifier(
+                table_name, TABLE_NAME_PATTERN, allow_reserved_words=False
+            )
     else:
-        assert check_table_name(table_name) is None  # type: ignore[func-returns-value]
+        assert is_safe_identifier(
+            table_name, TABLE_NAME_PATTERN, allow_reserved_words=False
+        )
+
+
+@pytest.mark.parametrize(
+    "column_name, expected",
+    [
+        # Valid column names
+        ("valid_column_name", None),
+        ("_another_valid_name123", None),
+        ("column_name_with_underscores", None),
+        ("COLUMNnameWITHCAPITALletters", None),
+        ("short", None),
+        ("a" * 63, None),  # Maximum length
+        ("name with-other*allowed:special/characters?!", None),
+        # Invalid identifiers - length
+        (
+            "a" * 64,
+            pytest.raises(
+                ValueError, match=r"Invalid SQL identifier.+max character number is 63"
+            ),
+        ),
+        # Invalid identifiers - malformed
+        (
+            "1invalid_start",
+            pytest.raises(ValueError, match=r"Malformed SQL identifier.+"),
+        ),
+        ("-invalid", pytest.raises(ValueError, match=r"Malformed SQL identifier.+")),
+        # Invalid identifiers - forbidden characters
+        (
+            'invalid"name',
+            pytest.raises(
+                ValueError,
+                match=r"Invalid SQL identifier.+contains forbidden character.+",
+            ),
+        ),
+        (
+            "invalid'name",
+            pytest.raises(
+                ValueError,
+                match=r"Invalid SQL identifier.+contains forbidden character.+",
+            ),
+        ),
+        (
+            "invalid`name",
+            pytest.raises(
+                ValueError,
+                match=r"Invalid SQL identifier.+contains forbidden character.+",
+            ),
+        ),
+        (
+            "invalid;name",
+            pytest.raises(
+                ValueError,
+                match=r"Invalid SQL identifier.+contains forbidden character.+",
+            ),
+        ),
+        (
+            "invalid--name",
+            pytest.raises(
+                ValueError,
+                match=r"Invalid SQL identifier.+contains forbidden character.+",
+            ),
+        ),
+        (
+            "invalid\\*name",
+            pytest.raises(
+                ValueError,
+                match=r"Invalid SQL identifier.+contains forbidden character.+",
+            ),
+        ),
+        (
+            "invalid*\\name",
+            pytest.raises(
+                ValueError,
+                match=r"Invalid SQL identifier.+contains forbidden character.+",
+            ),
+        ),
+        (
+            "invalid*/name",
+            pytest.raises(
+                ValueError,
+                match=r"Invalid SQL identifier.+contains forbidden character.+",
+            ),
+        ),
+        (
+            "invalid\\name",
+            pytest.raises(
+                ValueError,
+                match=r"Invalid SQL identifier.+contains forbidden character.+",
+            ),
+        ),
+        (
+            "invalid(name",
+            pytest.raises(
+                ValueError,
+                match=r"Invalid SQL identifier.+contains forbidden character.+",
+            ),
+        ),
+        (
+            "invalid)name",
+            pytest.raises(
+                ValueError,
+                match=r"Invalid SQL identifier.+contains forbidden character.+",
+            ),
+        ),
+        (
+            "invalid+name",
+            pytest.raises(
+                ValueError,
+                match=r"Invalid SQL identifier.+contains forbidden character.+",
+            ),
+        ),
+        (
+            "invalid=name",
+            pytest.raises(
+                ValueError,
+                match=r"Invalid SQL identifier.+contains forbidden character.+",
+            ),
+        ),
+        (
+            "yet\\another--invalid=name+with(many)forbidden*/characters",
+            pytest.raises(
+                ValueError,
+                match=r"Invalid SQL identifier.+contains forbidden character.+",
+            ),
+        ),
+    ],
+)
+def test_check_column_name_is_safe(column_name: str, expected: str) -> None:
+    if isinstance(expected, type(pytest.raises(ValueError))):
+        with expected:
+            is_safe_identifier(
+                column_name, COLUMN_NAME_PATTERN, allow_reserved_words=True
+            )
+    else:
+        assert is_safe_identifier(
+            column_name, COLUMN_NAME_PATTERN, allow_reserved_words=True
+        )
+
+
+@pytest.mark.parametrize("data_uri", ["sqlite_uri", "duckdb_uri", "postgres_uri"])
+@pytest.mark.parametrize("column_name", ["a", "a b", "a-b", "a:b", "a*b", "a/b", "Ab"])
+def test_can_query_with_valid_column_names(
+    data_uri: str, column_name: str, request: pytest.FixtureRequest
+) -> None:
+    table = pa.Table.from_arrays([[1, 2, 3]], [column_name])
+    structure = TableStructure.from_arrow_table(table)
+    data_source = DataSource(
+        management=Management.writable,
+        mimetype="application/x-tiled-sql-table",
+        structure_family=StructureFamily.table,
+        structure=structure,
+        assets=[],
+    )
+    data_uri = request.getfixturevalue(data_uri)
+    storage = parse_storage(data_uri)
+    register_storage(storage)
+    assert SQLAdapter.init_storage(data_source=data_source, storage=storage) is not None
+
+
+@pytest.mark.parametrize("data_uri", ["sqlite_uri", "duckdb_uri", "postgres_uri"])
+def test_reject_colliding_uppercase_column_names(
+    data_uri: str, request: pytest.FixtureRequest
+) -> None:
+    # Define a table and a storage
+    data_uri = request.getfixturevalue(data_uri)
+    storage = parse_storage(data_uri)
+    register_storage(storage)
+
+    # Create a table with colliding column names
+    table = pa.Table.from_arrays([[1, 2, 3], [4, 5, 6]], ["column_name", "COLUMN_NAME"])
+    structure = TableStructure.from_arrow_table(table)
+    data_source = DataSource(
+        management=Management.writable,
+        mimetype="application/x-tiled-sql-table",
+        structure_family=StructureFamily.table,
+        structure=structure,
+        parameters={"table_name": "table_name"},
+        assets=[],
+    )
+    with pytest.raises(ValueError, match=r"Column names must be unique.+"):
+        SQLAdapter.init_storage(data_source=data_source, storage=storage)
+
+    # Create a table with mixed cases in column names
+    table = pa.Table.from_arrays([[1, 2, 3], [4, 5, 6]], ["lower_case", "UPPER_CASE"])
+    structure = TableStructure.from_arrow_table(table)
+    data_source = DataSource(
+        management=Management.writable,
+        mimetype="application/x-tiled-sql-table",
+        structure_family=StructureFamily.table,
+        structure=structure,
+        parameters={"table_name": "table_name"},
+        assets=[],
+    )
+    data_source = SQLAdapter.init_storage(data_source=data_source, storage=storage)
+    assert data_source is not None
+
+    # Write to and read from the table
+    adapter = SQLAdapter(
+        data_source.assets[0].data_uri,
+        structure=data_source.structure,
+        table_name=data_source.parameters["table_name"],
+        dataset_id=data_source.parameters["dataset_id"],
+    )
+    adapter.append_partition(table, 0)
+    assert adapter.table_name == "table_name"
+    assert set(adapter.read().columns) == {"lower_case", "UPPER_CASE"}